+++ src/ray_tracing_library/include/Material.cuh
@@ -26,17 +26,17 @@ public:
     }
 
     __host__ __device__ static
-    Material getSpecular(const Color &albedo, const Color &specularColor, double roughness, double percentSpecular) {
+    Material getSpecular(const Color &albedo, const Color &specularColor, float roughness, float percentSpecular) {
         return {albedo, specularColor, roughness, percentSpecular};
     }
 
     __host__ __device__ static
-    Material getGlowing(const Color &albedo, const Color &emittedColor, double intensity) {
+    Material getGlowing(const Color &albedo, const Color &emittedColor, float intensity) {
         return {albedo, emittedColor * intensity};
     }
 
     __host__ __device__ static
-    Material getGlass(const Color &albedo, double refractionIdx) {
+    Material getGlass(const Color &albedo, float refractionIdx) {
         return {albedo, refractionIdx};
     }
 
@@ -44,40 +44,40 @@ private:
     Color _albedo{};
     Color _specularColor{};
     Color _emittedColor{};
-    double _percentSpecular{};
-    double _roughnessSquared{};
-    double _refractionIdx = 1.0;
+    float _percentSpecular{};
+    float _roughnessSquared{};
+    float _refractionIdx = 1.0f;
     bool _isRefractable{false};
 
 
     __host__ __device__
     explicit Material(const Color &albedo) : _albedo(albedo),
-                                             _specularColor({0, 0, 0}),
-                                             _emittedColor({0, 0, 0}),
-                                             _percentSpecular(0),
-                                             _roughnessSquared(0) {}
+                                             _specularColor({0.0f, 0.0f, 0.0f}),
+                                             _emittedColor({0.0f, 0.0f, 0.0f}),
+                                             _percentSpecular(0.0f),
+                                             _roughnessSquared(0.0f) {}
 
     __host__ __device__
     Material(const Color &albedo,
              const Color &specularColor,
-             double roughness,
-             double percentSpecular) : _albedo(albedo),
+             float roughness,
+             float percentSpecular) : _albedo(albedo),
                                        _specularColor(specularColor),
-                                       _emittedColor({0, 0, 0}),
+                                       _emittedColor({0.0f, 0.0f, 0.0f}),
                                        _percentSpecular(percentSpecular),
                                        _roughnessSquared(roughness * roughness) {}
 
     __host__ __device__
     Material(const Color &albedo, const Color &emittedColor) : _albedo(albedo),
-                                                               _specularColor({0, 0, 0}),
+                                                               _specularColor({0.0f, 0.0f, 0.0f}),
                                                                _emittedColor(emittedColor),
-                                                               _percentSpecular(0),
-                                                               _roughnessSquared(0) {}
+                                                               _percentSpecular(0.0f),
+                                                               _roughnessSquared(0.0f) {}
 
     __host__ __device__
-    Material(const Color &albedo, double refractionIdx) : _albedo(albedo),
-                                                          _specularColor({1, 1, 1}),
-                                                          _emittedColor({0, 0, 0}),
+    Material(const Color &albedo, float refractionIdx) : _albedo(albedo),
+                                                          _specularColor({1.0f, 1.0f, 1.0f}),
+                                                          _emittedColor({0.0f, 0.0f, 0.0f}),
                                                           _percentSpecular(1.0),
                                                           _roughnessSquared(0),
                                                           _refractionIdx(refractionIdx),
@@ -85,10 +85,10 @@ private:
 
 
     __device__
-    bool shouldDoReflection(const HitResult &hitRes, double refractionIdxRatio, Vec3 &rayDirNormalized,
+    bool shouldDoReflection(const HitResult &hitRes, float refractionIdxRatio, Vec3 &rayDirNormalized,
                             curandState *randState) const;
 
     __device__
-    void getSpecularResult(const HitResult &hitRes, Vec3 &diffuseDir, double specularChance, Vec3 &resDir,
+    void getSpecularResult(const HitResult &hitRes, Vec3 &diffuseDir, float specularChance, Vec3 &resDir,
                            Color &resColor, curandState *randState) const;
 };
