+++ src/ray_tracing_library/include/Sphere.cuh
@@ -13,18 +13,18 @@ class Sphere {
 public:
     __host__ __device__ Sphere() = default;
 
-    __host__ __device__ Sphere(const Point3 &center, double radius, const Material& mat)
-                                    : _center(center), _radius(radius), _material(mat) {}
+    __host__ __device__ Sphere(const Point3 &center, float radius, const Material& mat)
+            : _center(center), _radius(radius), _material(mat) {}
 
     __host__ __device__ void
-    getHitResult(const Ray &ray, double rootRes, HitResult &hitRes) const;
-    __host__ __device__ bool isHit(const Ray &ray, double tStart, double tEnd, double &rootRes) const;
+    getHitResult(const Ray &ray, float rootRes, HitResult &hitRes) const;
+    __host__ __device__ bool isHit(const Ray &ray, float tStart, float tEnd, float &rootRes) const;
 
     __host__ __device__ const Material &getMaterial() const;
 
 private:
     Point3 _center{};
-    double _radius{};
+    float _radius{};
     Material _material{};
 };
