+++ src/ray_tracing_library/src/Material.cu
@@ -13,7 +13,7 @@ __device__ void Material::getColorAndSecondaryRay(const HitResult &hitRes, curan
     Color resultColor{};
 
     Vec3 diffuseDir = normalize((hitRes.normal + randomUnitVec(randState)));
-    double specularChance = randomFloatCuda(randState);
+    float specularChance = randomFloatCuda(randState);
     if (specularChance < _percentSpecular) {
         getSpecularResult(hitRes, diffuseDir, specularChance, secondaryRayDir, resultColor, randState);
     } else {
@@ -26,9 +26,9 @@ __device__ void Material::getColorAndSecondaryRay(const HitResult &hitRes, curan
     resSecondaryRay = Ray(hitRes.hitPoint, secondaryRayDir);
 }
 
-__device__ void Material::getSpecularResult(const HitResult &hitRes, Vec3 &diffuseDir, double specularChance,
+__device__ void Material::getSpecularResult(const HitResult &hitRes, Vec3 &diffuseDir, float specularChance,
                                             Vec3 &resDir, Color &resColor, curandState *randState) const {
-    double refractionIdxRatio = hitRes.isOutwardsNormal ? 1.0 / _refractionIdx : _refractionIdx;
+    float refractionIdxRatio = hitRes.isOutwardsNormal ? 1.0f / _refractionIdx : _refractionIdx;
     Vec3 rayDirNormalized = normalize(hitRes.hittingRay.direction());
 
     Vec3 specularDir{};
@@ -43,9 +43,9 @@ __device__ void Material::getSpecularResult(const HitResult &hitRes, Vec3 &diffu
     resColor = alphaBlending(_specularColor, _albedo, specularChance);
 }
 
-__device__ bool Material::shouldDoReflection(const HitResult &hitRes, double refractionIdxRatio, Vec3 &rayDirNormalized,
+__device__ bool Material::shouldDoReflection(const HitResult &hitRes, float refractionIdxRatio, Vec3 &rayDirNormalized,
                                              curandState *randState) const {
-    double cosTheta = fmin(dot(-rayDirNormalized, hitRes.normal), 1.0);
+    float cosTheta = fminf(dot(-rayDirNormalized, hitRes.normal), 1.0);
     bool cannotRefract = cannotRefractBySnellsLaw(cosTheta, refractionIdxRatio);
     bool reflectApprox = reflectSchlickApproxForFrensel(cosTheta, refractionIdxRatio) > randomFloatCuda(randState);
     bool doReflection = !_isRefractable || cannotRefract || reflectApprox;
